Split IO part of metacling/rootclingTCling.cxx into utils/rootclingIO.cxx; the rest depends is meta not cling.
